finished refactoring and implements classes for each phase. Each phase runs a set of rules, which can be invoked with the Phase's apply_rules() method
